split entries classes into two files and move.
