+++ pacman/minirig/place_and_route/route/ner.py
@@ -19,7 +19,7 @@ from pacman.operations.rig_algorithms.utils import longest_dimension_first
 from pacman.exceptions import MachineHasDisconnectedSubregion
 
 from pacman.operations.rig_algorithms.links import Links
-from pacman.minirig.routing_table.entries import Routes
+from pacman.operations.rig_algorithms.routes import Routes
 
 from pacman.minirig.place_and_route.routing_tree import RoutingTree
 from pacman.model.graphs import (
