+++ pacman/operations/router_compressors/mundys_router_compressor/routing_table_condenser.py
@@ -6,7 +6,8 @@ from spinn_machine import MulticastRoutingEntry
 from pacman.model.routing_tables import (
     MulticastRoutingTable, MulticastRoutingTables)
 from pacman.exceptions import PacmanElementAllocationException
-from pacman.minirig.routing_table.entries import (RoutingTableEntry, Routes)
+from pacman.operations.rig_algorithms.routing_table_entry import RoutingTableEntry
+from pacman.operations.rig_algorithms.routes import Routes
 from pacman.minirig.routing_table import ordered_covering as rigs_compressor
 
 logger = logging.getLogger(__name__)
